move -dl to repo group
