+++ bot.py
@@ -26,14 +26,13 @@ logger = logging.getLogger('main')
 
 extensions: list = [
     'core.background',
-    'cogs.base.user',
-    'cogs.base.org',
-    'cogs.base.repo',
-    'cogs.numbered.pr',
-    'cogs.numbered.issue',
-    'cogs.download',
-    'cogs.lines',
-    'cogs.info',
+    'cogs.github.base.user',
+    'cogs.github.base.org',
+    'cogs.github.base.repo',
+    'cogs.github.numbered.pr',
+    'cogs.github.numbered.issue',
+    'cogs.github.other.lines',
+    'cogs.github.other.info',
     'cogs.help',
     'cogs.config',
     'cogs.debug',
